Move details date formatting to utils
