+++ client/src/utils/vehicleDates.js
@@ -42,3 +42,14 @@ export function addExpiryFlags(vehicle) {
     expiring_revision: revision.expiring,
   };
 }
+
+export function formatDateForInput(dateString) {
+  const date = parseDate(dateString);
+  if (!date) return "";
+
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+
+  return `${year}-${month}-${day}`;
+}
